[IMP] start add shopinvader_product_brand
