+++ shopinvader_search_engine/models/__init__.py
@@ -1,4 +1,5 @@
 from . import shopinvader_backend
+from . import shopinvader_se_binding
 from . import shopinvader_variant
 from . import shopinvader_category
 from . import se_index
